*) Changes from Dre: - Changed: some indexes in tables based on queries. - Added: '/object copy' command to copy a targeted object. - Fixed: LanguageMgr.GetTranslation don't fail when "client.Account" is null. - Changed: No more useless cast when access to GuildMgr.SocialWindowMember indexer.

+++ GameServer/commands/playercommands/guild.cs
@@ -1438,7 +1438,7 @@ namespace DOL.GS.Commands
 												{
 													GuildMgr.SocialWindowMember member = guildMembers[player.Name];
 													member.UpdateMember(player);
-													string key = member[(int)index];
+													string key = member[index];
 													if (sortedList.ContainsKey(key))
 														key += sortedList.Count.ToString();
 
@@ -1463,7 +1463,7 @@ namespace DOL.GS.Commands
 													//Make sure that since they are offline they get the offline flag!
 													member.GroupSize = "0";
 												//Add based on the new index
-												string key = member[(int)index];
+												string key = member[index];
 												if (sortedList.ContainsKey(key))
 												{
 													key += sortedList.Count.ToString();
